requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Session==0.4.0
gunicorn==20.1.0
idna==3.3
imageio==2.13.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
paramiko==2.8.1
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.6
python-dotenv==0.19.2
PyWavelets==1.2.0
requests==2.26.0
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
